requirements_venv310.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
bce-python-sdk==0.9.69
certifi==2026.2.25
chardet==7.4.2
charset-normalizer==3.4.7
click==8.3.2
colorlog==6.10.1
comm==0.2.3
crc32c==2.8
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
future==1.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
imagesize==2.0.0
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
modelscope==1.35.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.0.2.14
nvidia-cuda-cccl==13.0.85
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.88
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.1
nvidia-nccl-cu13==2.28.3
nvidia-nvjitlink==13.0.88
nvidia-nvtx==13.0.85
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opt-einsum==3.3.0
packaging==26.0
paddleocr==3.4.0
paddlepaddle-gpu==3.3.0
paddlex==3.4.3
pandas==2.3.3
parso==0.8.6
pdf2image==1.17.0
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
pypdfium2==5.7.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.33.1
rich==15.0.0
ruamel.yaml==0.19.1
safetensors==0.7.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
